+++ design-system/tokens/scripts/style-dictionary/sort-tokens.tsx
@@ -5,62 +5,7 @@
 
 import type { TransformedToken } from 'style-dictionary';
 
-/**
- * Create specific rules for ordering tokens based on their root path and subpath.
- */
-export const tokenOrder = [
-  {
-    path: 'color',
-    subpaths: [
-      'text',
-      'link',
-      'icon',
-      'border',
-      'background',
-      'blanket',
-      'interaction',
-      'skeleton',
-      'chart',
-      // deleted ↓
-      'accent',
-      'iconBorder',
-      'overlay',
-    ],
-  },
-  {
-    path: 'elevation',
-    subpaths: ['surface', 'shadow'],
-  },
-  {
-    path: 'opacity',
-    subpaths: [],
-  },
-  { path: 'shadow', subpaths: ['card', 'overlay'] }, // Deleted
-  {
-    path: 'utility',
-    subpaths: [],
-  },
-  {
-    path: 'shape',
-    subpaths: ['radius', 'width'],
-  },
-  {
-    path: 'space',
-    subpaths: [],
-  },
-  {
-    path: 'font',
-    subpaths: ['family', 'size', 'weight', 'lineHeight'],
-  },
-  {
-    path: 'typography',
-    subpaths: [],
-  },
-  {
-    path: 'value', // Legacy palette
-    subpaths: [],
-  },
-];
+import { tokenOrder } from '../../src/utils/token-order';
 
 const removeDefaultPaths = (token: TransformedToken) =>
   token.path.filter((p) => p !== '[default]');