+++ design-system/tokens/src/index.tsx
@@ -8,7 +8,16 @@ export {
 } from './set-global-theme';
 export type { ThemeState } from './set-global-theme';
 export type { CSSToken } from './artifacts/token-names';
+export type { ActiveTokens } from './artifacts/types';
 export type { ThemeColorModes, Themes, ThemeIds } from './theme-config';
+export type {
+  Groups,
+  OpacityToken,
+  PaintToken,
+  RawToken,
+  ShadowToken,
+  SpacingToken,
+} from './types';
 export { default as themeConfig } from './theme-config';
 export { useThemeObserver } from './use-theme-observer';
 export { ThemeMutationObserver } from './theme-mutation-observer';