+++ design-system/tokens/src/utils/color-detection.tsx
@@ -1,4 +1,3 @@
-import { token } from '../index';
 import { ShadowToken } from '../types';
 
 export const hexToRGBAValues = (hex: string) => {
@@ -10,74 +9,6 @@ export const hexToRGBAValues = (hex: string) => {
     b: parseInt(hexColor.slice(4, 6), 16),
     a: parseFloat((parseInt(hexColor.slice(6, 8), 16) / 255).toFixed(2)),
   };
-};
-
-export const hexToRGBA = (hex: string) => {
-  const { r, g, b, a } = hexToRGBAValues(hex);
-
-  return `rgb${a ? 'a' : ''}(${r},${g},${b}${a ? `,${a}` : ''})`;
-};
-
-export const getLuminance = ({
-  r,
-  g,
-  b,
-}: {
-  r: number;
-  b: number;
-  g: number;
-}) => (r * 299 + g * 587 + b * 114) / 1000;
-
-/**
- * Returns an accessible hard-coded text color based on the color contrast with
- * the background.
- *
- * @param hex - The Hex color code of the background
- * @param [opts.hardcodedSurface] - If set, a design token will be returned instead
- * of a hard-coded color. This is to support more transparent backgrounds
- * to allow the text to invert colors depending on the current theme's surface color.
- */
-export const getTextColorForBackground = (
-  hex: string,
-  opts?: { hardcodedSurface?: 'light' | 'dark' },
-): string => {
-  const { r, g, b, a } = hexToRGBAValues(hex);
-  const lum = getLuminance({ r, g, b });
-  const alphaLimit = 0.42;
-
-  const alphaConditionsPerSurface: {
-    light: boolean;
-    dark: boolean;
-  } = {
-    light: a < alphaLimit,
-    dark: a > alphaLimit,
-  };
-
-  const alphaLimitExceeded =
-    opts?.hardcodedSurface && alphaConditionsPerSurface[opts.hardcodedSurface];
-
-  if (!opts?.hardcodedSurface && a < alphaLimit) {
-    // This color is transparent, so the text will mainly cast onto the surface behind.
-    // Needs to use tokens otherwise Dark mode would cause black text on black surface
-    return token('color.text', 'black');
-  }
-
-  return (lum > 150 && !a) || (a && alphaLimitExceeded) ? 'black' : 'white';
-};
-
-/**
- * Returns a border if determined to be required based on the color contrast with
- * the background.
- *
- * @param hex - The Hex color code of the background
- */
-export const getBorderForBackground = (hex: string) => {
-  const { r, g, b, a } = hexToRGBAValues(hex);
-  const lum = getLuminance({ r, g, b });
-
-  return lum > 240 || a < 0.2
-    ? `1px solid ${token('color.border', '#091E4224')}`
-    : undefined;
 };
 
 /**