+++ editor/editor-core/src/plugins/date/__tests__/integration/escape-keydown.ts
@@ -27,9 +27,7 @@ const adf = {
   ],
 };
 
-// TODO: https://product-fabric.atlassian.net/browse/ED-16966
 runEscapeKeydownSuite({
   adf,
   openMenu: clickOnDate,
-  skip: ['*'],
 });