+++ linking-platform/linking-types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/linking-types
 
+## 0.1.1
+
+### Patch Changes
+
+- [`47412a09e69`](https://bitbucket.org/atlassian/atlassian-frontend/commits/47412a09e69) - Changed export way for SmartLinkActionType enum
+
 ## 0.1.0
 
 ### Minor Changes